Moved my event form submission
