+++ components/event_card.templ
@@ -11,7 +11,7 @@ templ EventCard(eventCard models.EventCardModel, eventImageDir *string, showStat
 	{{ imageUrl := eventimage.GetEventImageUrl(eventCard.Id, "card", eventImageDir) }}
 	<a
 		if eventCard.Status == models.EventStatusDraft || eventCard.Status == models.EventStatusSubmitted || eventCard.Status == models.EventStatusArchived {
-			href={ templ.URL(fmt.Sprintf("/my-events/new/%s", eventCard.Id)) }
+			href={ templ.URL(fmt.Sprintf("/profile/new/%s", eventCard.Id)) }
 		} else {
 			if pulje != "" {
 				href={ templ.URL(fmt.Sprintf("/event/%s?pulje=%s", eventCard.Id, pulje)) }
