+++ components/formsubmission/contact_info.templ
@@ -111,7 +111,7 @@ templ contactInfo(eventId string, name string, email string, phone string) {
 					name="name"
 					data-bind="name"
 					data-indicator="toastName"
-					data-on:change={ fmt.Sprintf("@put('/my-events/api/new/%s/name')", eventId) }
+					data-on:change={ fmt.Sprintf("@put('/profile/api/new/%s/name')", eventId) }
 					class="input"
 					placeholder="Ditt navn"
 					required
@@ -128,7 +128,7 @@ templ contactInfo(eventId string, name string, email string, phone string) {
 					type="email"
 					data-bind="email"
 					data-indicator="toastEmail"
-					data-on:change={ fmt.Sprintf("@put('/my-events/api/new/%s/email')", eventId) }
+					data-on:change={ fmt.Sprintf("@put('/profile/api/new/%s/email')", eventId) }
 					name="email"
 					class="input"
 					placeholder="Din e-post"
@@ -143,7 +143,7 @@ templ contactInfo(eventId string, name string, email string, phone string) {
 					data-signals={ dataSignals(map[string]string{"phone": phone}) }
 					data-bind="phone"
 					data-indicator="toastPhone"
-					data-on:change={ fmt.Sprintf("@put('/my-events/api/new/%s/phone')", eventId) }
+					data-on:change={ fmt.Sprintf("@put('/profile/api/new/%s/phone')", eventId) }
 					name="phone"
 					class="input"
 					placeholder="Ditt telefonnummer"
