+++ components/formsubmission/puljefordeling.templ
@@ -185,7 +185,7 @@ templ puljeRow(eventId string, pulje models.PuljeRow, db *sql.DB) {
 						}
 						data-bind={ fmt.Sprintf("isInPulje%s", pulje.ID) }
 						data-indicator={ fmt.Sprintf("toastPuljeMembership%s", pulje.ID) }
-						data-on:change={ fmt.Sprintf("@put('/my-events/api/new/%s/event-in-pulje/%s')", eventId, pulje.ID) }
+						data-on:change={ fmt.Sprintf("@put('/profile/api/new/%s/event-in-pulje/%s')", eventId, pulje.ID) }
 						type="checkbox"
 						name="pulje"
 						class="checkbox input"
@@ -202,7 +202,7 @@ templ puljeRow(eventId string, pulje models.PuljeRow, db *sql.DB) {
 						}
 						data-bind={ fmt.Sprintf("isPublished%s", pulje.ID) }
 						data-indicator={ fmt.Sprintf("toastPuljePublished%s", pulje.ID) }
-						data-on:change={ fmt.Sprintf("@put('/my-events/api/new/%s/is-published/%s')", eventId, pulje.ID) }
+						data-on:change={ fmt.Sprintf("@put('/profile/api/new/%s/is-published/%s')", eventId, pulje.ID) }
 						type="checkbox"
 						name="pulje"
 						class="checkbox input"
