+++ components/formsubmission/statusCard.templ
@@ -54,7 +54,7 @@ templ statusCard(event *models.Event, isAdmin bool) {
 				data-signals={ dataSignals(map[string]models.EventStatus{"status": event.Status}) }
 				data-bind="status"
 				data-indicator="toastStatus"
-				data-on:change={ fmt.Sprintf("@put('/my-events/api/new/%s/status')", event.ID) }
+				data-on:change={ fmt.Sprintf("@put('/profile/api/new/%s/status')", event.ID) }
 				required
 			>
 				if isAdmin {
