+++ components/profile/my_events.templ
@@ -6,7 +6,7 @@ import (
 )
 
 templ singleEvent(event models.EventCardModel) {
-	{{ eventLink := "/my-events/new/" + event.Id }}
+	{{ eventLink := "/profile/new/" + event.Id }}
 	if event.Status == models.EventStatusSubmitted || event.Status == models.EventStatusApproved || event.Status == models.EventStatusPublished {
 		{{ eventLink = "/event/" + event.Id }}
 	}
@@ -21,7 +21,7 @@ templ singleEvent(event models.EventCardModel) {
 }
 
 templ MyEvents(events []models.EventCardModel) {
- <style>
+	<style>
     .submit-event-message {
         display: flex;
         width: 100%;
@@ -53,7 +53,8 @@ templ MyEvents(events []models.EventCardModel) {
 				@singleEvent(event)
 			}
 		}
-		<form action="/my-events/api/create" method="post" class="submit-event-message">
+		<form action="/profile/api/create" method="post" class="submit-event-message">
+			profile
 			<p>
 				Ønsker du å arrangere noe under festivalen? Dørene er fortsatt åpne for nye arrangementer, la oss høre fra
 				deg!
