--- pages/myprofile/myevents/myevents_page.templ
@@ -1,110 +1,0 @@
-package myevents
-
-import (
-	"database/sql"
-	"fmt"
-	"github.com/Regncon/conorganizer/components"
-	"github.com/Regncon/conorganizer/models"
-	"github.com/Regncon/conorganizer/service/userctx"
-	"log/slog"
-)
-
-func GetEventsByExternalID(db *sql.DB, externalID string, logger *slog.Logger) []models.EventCardModel {
-	var events []models.EventCardModel
-
-	userID, userIDErr := userctx.GetUserIDFromExternalID(externalID, db, logger)
-	if userIDErr != nil {
-		logger.Error(fmt.Errorf("failed to resolve user_id for external_id %q: %w", externalID, userIDErr).Error())
-		return events
-	}
-
-	eventsQuery := "SELECT id, title, intro, status, system, host_name,beginner_friendly, event_type, age_group, event_runtime, can_be_run_in_english FROM events WHERE user_id = ?"
-	rows, eventsQueryErr := db.Query(eventsQuery, userID)
-	if eventsQueryErr != nil {
-		logger.Error(fmt.Errorf("failed to query events for external_id %q: %w", externalID, eventsQueryErr).Error())
-		return events
-	}
-	defer rows.Close()
-
-	for rows.Next() {
-		var event models.EventCardModel
-		if scanErr := rows.Scan(&event.Id, &event.Title, &event.Intro, &event.Status, &event.System, &event.HostName, &event.BeginnerFriendly, &event.EventType, &event.AgeGroup, &event.Runtime, &event.CanBeRunInEnglish); scanErr != nil {
-			logger.Error(fmt.Errorf("failed to scan event row for external_id %q: %w", externalID, scanErr).Error())
-			return events
-		}
-		events = append(events, event)
-	}
-	logger.Debug("Found events for user", "event_count", len(events), "external_id", externalID, "user_id", userID)
-	return events
-}
-
-templ myEventsPage(externalID string, db *sql.DB, eventImageDir *string, logger *slog.Logger) {
-	<div id="my-events-wrapper">
-		@myEventsPageContent(externalID, db, eventImageDir, logger)
-	</div>
-}
-
-templ myEventsPageContent(externalID string, db *sql.DB, eventImageDir *string, logger *slog.Logger) {
-	@components.Breadcrumbs([]components.BreadcrumbPath{
-		{Name: "Hjem", Url: "/"},
-		{Name: "Mine arrangementer", Url: ""},
-	})
-	<div class="page-content-container">
-		<h1 class="page-heading">Mine arrangementer</h1>
-		<div class="my-event-grid">
-			@events(externalID, db, eventImageDir, logger)
-		</div>
-	</div>
-	<style>
-		.my-event-grid {
-			display: grid;
-			grid-template-columns: repeat(
-				auto-fill,
-				minmax(var(--mobile-min-width), 1fr)
-			);
-			gap: 1rem;
-		}
-	</style>
-}
-
-templ events(externalID string, db *sql.DB, eventImageDir *string, logger *slog.Logger) {
-	{{ events := GetEventsByExternalID(db, externalID, logger) }}
-	if len(events) == 0 {
-		@createEvent()
-	}
-	if len(events) > 0 {
-		for _, event := range events {
-			@components.EventCard(event, eventImageDir, true, "")
-		}
-		@createEvent()
-	}
-}
-
-templ createEvent() {
-	<form action="/my-events/api/create" method="post">
-		<button
-			type="submit"
-			style="
-                width: 100%;
-                height: 100%;
-                aspect-ratio: 41 / 34;
-		        width: var(--mobile-responsive-min-width);
-		        background-color: var(--bg-surface);
-		        border-radius: var(--border-radius-1x);
-		        display: flex;
-		        flex-direction: column;
-		        text-decoration: none;
-		        color: var(--color-primary-text);
-		        box-sizing: border-box;
-		        place-items: center;
-		        place-content: center;
-		        padding-block: 2rem;
-		        text-align: center;
-                border-color: var(--bg-item-border);
-		        "
-		>
-			<h2 style="margin: 0">Legg til nytt arrangement</h2>
-			<h1 style="margin: 0; font-size: 3rem">➕</h1>
-		</button>
-	</form>
-}
