--- pages/myprofile/myevents/nats.go
@@ -1,35 +1,0 @@
-package myevents
-
-import (
-	"context"
-	"fmt"
-	"time"
-
-	"github.com/delaneyj/toolbelt/embeddednats"
-	"github.com/nats-io/nats.go/jetstream"
-)
-
-func SetupNats(ns *embeddednats.Server) (jetstream.KeyValue, error) {
-	nc, err := ns.Client()
-	if err != nil {
-		return nil, fmt.Errorf("error creating nats client: %w", err)
-	}
-
-	js, err := jetstream.New(nc)
-	if err != nil {
-		return nil, fmt.Errorf("error creating jetstream client: %w", err)
-	}
-
-	kv, err := js.CreateOrUpdateKeyValue(context.Background(), jetstream.KeyValueConfig{
-		Bucket:      "events",
-		Description: "Regncon Event Store",
-		Compression: true,
-		TTL:         time.Hour,
-		MaxBytes:    16 * 1024 * 1024,
-	})
-
-	if err != nil {
-		return nil, fmt.Errorf("error creating key value: %w", err)
-	}
-	return kv, nil
-}
