--- pages/myprofile/myprofile.go
@@ -1,75 +1,0 @@
-package myprofile
-
-import (
-	"database/sql"
-	"fmt"
-	"log/slog"
-	"net/http"
-
-	"github.com/Regncon/conorganizer/layouts"
-	"github.com/Regncon/conorganizer/service/checkIn"
-	"github.com/Regncon/conorganizer/service/userctx"
-	"github.com/delaneyj/toolbelt/embeddednats"
-	"github.com/go-chi/chi/v5"
-	"github.com/gorilla/sessions"
-)
-
-func SetupMyProfileRoute(router chi.Router, store sessions.Store, ns *embeddednats.Server, db *sql.DB, logger *slog.Logger) error {
-	baseLogger := logger
-	logger = logger.With("component", "my_profile")
-
-	router.Route("/my-profile", func(ticketRouter chi.Router) {
-		ticketRouter.Get("/", func(w http.ResponseWriter, r *http.Request) {
-			var ctx = r.Context()
-			if err := layouts.Base(
-				"Min profil",
-				userctx.GetUserRequestInfo(ctx),
-				myProfile(),
-			).Render(ctx, w); err != nil {
-				logger.Error(fmt.Errorf("failed to render profile page: %w", err).Error())
-			}
-
-		})
-
-		ticketRouter.Get("/my-tickets", func(w http.ResponseWriter, r *http.Request) {
-			var ctx = r.Context()
-
-			tickets, ticketsErr := checkIn.GetTicketsFromCheckIn(baseLogger, userctx.GetUserRequestInfo(ctx).Email)
-
-			if ticketsErr != nil {
-				logger.Error(fmt.Errorf("failed to get tickets from check-in: %w", ticketsErr).Error())
-
-				if err := layouts.Base(
-					"Ingen billetter",
-					userctx.GetUserRequestInfo(ctx),
-					noTickets(),
-				).Render(ctx, w); err != nil {
-					logger.Error(fmt.Errorf("failed to render no tickets page: %w", err).Error())
-				}
-				return
-			}
-
-			if len(tickets) == 0 {
-				if err := layouts.Base(
-					"Ingen billetter",
-					userctx.GetUserRequestInfo(ctx),
-					noTickets(),
-				).Render(ctx, w); err != nil {
-					logger.Error(fmt.Errorf("failed to render no tickets page: %w", err).Error())
-				}
-				return
-			}
-
-			if err := layouts.Base(
-				"Mine billetter",
-				userctx.GetUserRequestInfo(ctx),
-				myTickets(),
-			).Render(ctx, w); err != nil {
-				logger.Error(fmt.Errorf("failed to render tickets page: %w", err).Error())
-			}
-
-		})
-	})
-
-	return nil
-}
