+++ pages/profile/newevent/new_page.templ
@@ -1,4 +1,4 @@
-package newEvent
+package newevent
 
 import (
 	"context"
@@ -20,9 +20,9 @@ templ NewEventFormPage(eventId string, userId string, ctx context.Context, db *s
 
 templ NewEventFormPageContent(eventId string, userId string, ctx context.Context, db *sql.DB, eventImageDir *string, logger *slog.Logger) {
 	{{
-	isAdmin := authctx.GetAdminFromUserToken(ctx)
+		isAdmin := authctx.GetAdminFromUserToken(ctx)
 		eventData, eventErr := eventservice.GetEventById(eventId, db)
-	isRemovedPreviousNext := components.PreviousNext{IsRemoved: true}
+		isRemovedPreviousNext := components.PreviousNext{IsRemoved: true}
 		puljerForEvent, puljerErr := eventservice.GetPujerForEvent(eventId, db)
 	}}
 	if eventErr != nil {
@@ -53,7 +53,7 @@ templ NewEventFormPageContent(eventId string, userId string, ctx context.Context
 		<div style="display: flex;">
 			<form
 				class="formsubmission-wrapper page-content-container"
-				action={ "/my-events/api/new/" + eventData.ID + "/submit" }
+				action={ "/profile/api/new/" + eventData.ID + "/submit" }
 				method="post"
 			>
 				<h1 class="page-heading">Nytt arrangement</h1>
