+++ plans/form-textarea-trim-normalization.md
@@ -8,19 +8,19 @@ Standardize Datastar textarea behavior so all three fields trim whitespace on in
 - Keep `data-signals:notes={ fmt.Sprintf("'%s'", strings.TrimSpace(notes)) }`.
 - Keep `data-bind:notes`.
 - Keep `data-on-input="$notes = $notes.trim()"`.
-- Keep existing `data-on-change={ datastar.PutSSE("/my-events/api/new/%s/notes", eventId) }`.
+- Keep existing `data-on-change={ datastar.PutSSE("/profile/api/new/%s/notes", eventId) }`.
 
 2. Update `intro` in `components/formsubmission/about_event.templ`.
 - Add `data-signals:intro={ fmt.Sprintf("'%s'", strings.TrimSpace(intro)) }`.
 - Change `data-bind="intro"` to `data-bind:intro`.
 - Add `data-on-input="$intro = $intro.trim()"`.
-- Keep existing `data-on-change={ datastar.PutSSE("/my-events/api/new/%s/intro", eventId) }`.
+- Keep existing `data-on-change={ datastar.PutSSE("/profile/api/new/%s/intro", eventId) }`.
 
 3. Update `description` in `components/formsubmission/about_event.templ`.
 - Add `data-signals:description={ fmt.Sprintf("'%s'", strings.TrimSpace(description)) }`.
 - Change `data-bind="description"` to `data-bind:description`.
 - Add `data-on-input="$description = $description.trim()"`.
-- Keep existing `data-on-change={ datastar.PutSSE("/my-events/api/new/%s/description", eventId) }`.
+- Keep existing `data-on-change={ datastar.PutSSE("/profile/api/new/%s/description", eventId) }`.
 
 4. Imports.
 - Ensure `about_event.templ` includes `strings`.
@@ -31,9 +31,9 @@ Standardize Datastar textarea behavior so all three fields trim whitespace on in
 2. Verify internal spaces remain unchanged.
 3. Verify whitespace-only input becomes empty string.
 4. Blur/change each field and verify trimmed values are sent to:
-- `/my-events/api/new/{id}/notes`
-- `/my-events/api/new/{id}/intro`
-- `/my-events/api/new/{id}/description`
+- `/profile/api/new/{id}/notes`
+- `/profile/api/new/{id}/intro`
+- `/profile/api/new/{id}/description`
 5. Reload and verify all three textareas initialize with trimmed values.
 
 ## Assumptions
