+++ schema.sql
@@ -150,47 +150,6 @@ FROM
     relation_billettholdere_users AS bu
     LEFT JOIN users AS u ON u.id = bu.user_id
 /* v_get_user_billettholder(user_id,external_id,user_email,user_is_admin,user_inserted_at,billettholder_id,billettholder_user_id,billettholder_user_inserted_at) */;
-CREATE VIEW v_events_by_pulje_active AS
-SELECT
-    e.id AS id,
-    e.title,
-    e.intro,
-    e.description,
-    e.system,
-    e.event_type,
-    e.age_group,
-    e.event_runtime,
-    e.host_name,
-    e.user_id,
-    e.email,
-    e.phone_number,
-    e.max_players,
-    e.beginner_friendly,
-    e.can_be_run_in_english,
-    e.notes,
-    e.status,
-    e.created_at,
-    ep.is_published AS is_published,
-    ep.pulje_id,
-    ep.room_id,
-    r.room_number,
-    r.name AS room_name,
-    r.floor AS room_floor,
-    r.max_concurrent_games AS room_max_concurrent_games,
-    r.notes AS room_notes,
-    r.is_disabled AS room_is_disabled,
-    p.name AS pulje_name,
-    p.start_at AS pulje_start_at,
-    p.end_at AS pulje_end_at
-FROM
-    events e
-    INNER JOIN relation_event_puljer ep ON ep.event_id = e.id
-    INNER JOIN puljer p ON p.id = ep.pulje_id
-    LEFT JOIN rooms r ON r.id = ep.room_id
-WHERE
-    e.status = 'Godkjent'
-    AND ep.is_in_pulje = 1
-/* v_events_by_pulje_active(id,title,intro,description,system,event_type,age_group,event_runtime,host_name,user_id,email,phone_number,max_players,beginner_friendly,can_be_run_in_english,notes,status,created_at,is_published,pulje_id,room_id,room_number,room_name,room_floor,room_max_concurrent_games,room_notes,room_is_disabled,pulje_name,pulje_start_at,pulje_end_at) */;
 CREATE VIEW v_event_summary AS
 SELECT
     e.id,
@@ -266,3 +225,43 @@ CREATE TABLE "puljer"(
   end_at TEXT NOT NULL,
   FOREIGN KEY(status) REFERENCES pulje_statuses(status) ON UPDATE CASCADE
 ) STRICT;
+CREATE VIEW v_events_by_pulje_active AS
+SELECT
+    e.id AS id,
+    e.title,
+    e.intro,
+    e.description,
+    e.system,
+    e.event_type,
+    e.age_group,
+    e.event_runtime,
+    e.host_name,
+    e.user_id,
+    e.email,
+    e.phone_number,
+    e.max_players,
+    e.beginner_friendly,
+    e.can_be_run_in_english,
+    e.notes,
+    e.status,
+    e.created_at,
+    ep.is_published AS is_published,
+    ep.pulje_id,
+    ep.room_id,
+    r.room_number,
+    r.name AS room_name,
+    r.floor AS room_floor,
+    r.max_concurrent_games AS room_max_concurrent_games,
+    r.notes AS room_notes,
+    r.is_disabled AS room_is_disabled,
+    p.name AS pulje_name,
+    p.start_at AS pulje_start_at,
+    p.end_at AS pulje_end_at
+FROM events e
+INNER JOIN relation_event_puljer ep ON ep.event_id = e.id
+INNER JOIN puljer p ON p.id = ep.pulje_id
+LEFT JOIN rooms r ON r.id = ep.room_id
+WHERE
+    e.status = 'Annonsert'
+    AND ep.is_in_pulje = 1
+/* v_events_by_pulje_active(id,title,intro,description,system,event_type,age_group,event_runtime,host_name,user_id,email,phone_number,max_players,beginner_friendly,can_be_run_in_english,notes,status,created_at,is_published,pulje_id,room_id,room_number,room_name,room_floor,room_max_concurrent_games,room_notes,room_is_disabled,pulje_name,pulje_start_at,pulje_end_at) */;
